imporant fix on the iterator, the check, in the while if statmeent was the wrong way round, leading to accessing of an element and checking its size when it didn't exist, corrupting memory

+++ src/data_structures/APR/APRAccess.hpp
@@ -949,7 +949,7 @@ public:
 
             gap_map.data[level][offset_pc_data].resize(1);
 
-            for (int i = global_begin; i < (global_begin + number_gaps) ; ++i) {
+            for (uint64_t i = global_begin; i < (global_begin + number_gaps) ; ++i) {
                 gap.y_end = map_data.y_end[i];
                 gap.global_index_begin = map_data.global_index[i];
